# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  /api/v1/auth/token:
    post:
      tags:
        - Управление токенами доступа
      summary: Получение токена доступа
      operationId: getNewAccessToken
      requestBody:
        description: Запрос на получение токена доступа по токену обновления
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshJwtRequest'
            examples:
              Запрос на получение токена доступа по токену обновления:
                description: Запрос на получение токена доступа по токену обновления
                value:
                  requestId: 123e4567-e89b-12d3-a456-426614174000
                  refreshToken: SuperMan
        required: true
      responses:
        '200':
          description: Успешная обработка запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtResponse'
              examples:
                Ответ с токенами доступа и обновления:
                  description: Ответ с токенами доступа и обновления
                  value:
                    operationStatus: COMPLETED
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    accessToken: >-
                      eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJTdXBlck1hbjQiLCJleHAiOjE3NjAzMTA1NzcsInJvbGVzIjpbIlVTRVIiXX0.0jcf_wlE5r49nMnlvL5d_HFi0w2f-ssCsPlpcTWwUZTjbfCkzVi8KHmKPoemhoWijCSwnFqMu9RuPST2YC2NxA
                    refreshToken: >-
                      eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJTdXBlck1hbjQiLCJleHAiOjE3NjI5MDIyNzd9.wFx7TLzzErJ9unezD2Znu7JC0z-f_PQWuHmASsfgDGtCTcz98-TD_Ncvw5qeV_lwv3FrasiuBnByqFCArbBAuA
                    type: Bearer
                Ответ с токенами доступа:
                  description: Ответ с токенами доступа
                  value:
                    operationStatus: COMPLETED
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    accessToken: >-
                      eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJTdXBlck1hbjQiLCJleHAiOjE3NjAzMTA1NzcsInJvbGVzIjpbIlVTRVIiXX0.0jcf_wlE5r49nMnlvL5d_HFi0w2f-ssCsPlpcTWwUZTjbfCkzVi8KHmKPoemhoWijCSwnFqMu9RuPST2YC2NxA
                    type: Bearer
        '400':
          description: >-
            Запрос не соответствует заявленной схеме данных и не может быть
            обработан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке валидации:
                  description: Ответ с сообщением об ошибке валидации
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Нарушение формата данных
        '403':
          description: Обработка закончилась с ошибкой авторизации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке авторизации:
                  description: Ответ с сообщением об ошибке авторизации
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Неправильный пароль
        '404':
          description: Обработка закончилась с ошибкой поиска
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке поиска:
                  description: Ответ с сообщением об ошибке поиска
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Пользователь не найден
        '500':
          description: Обработка закончилась внутренней ошибкой сервиса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением о внутренней ошибкой сервиса:
                  description: Ответ с сообщением о внутренней ошибкой сервиса
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Внутренняя ошибка сервиса
  /api/v1/auth/refresh:
    post:
      tags:
        - Управление токенами доступа
      summary: Получение токена доступа и нового токена обновления
      operationId: getNewRefreshToken
      requestBody:
        description: >-
          Запрос на получение токена доступа и нового токена обновления по
          токену обновления
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshJwtRequest'
            examples:
              Запрос на получение токена доступа и нового токена обновления по токену обновления:
                description: >-
                  Запрос на получение токена доступа и нового токена обновления
                  по токену обновления
                value:
                  requestId: 123e4567-e89b-12d3-a456-426614174000
                  refreshToken: SuperMan
        required: true
      responses:
        '200':
          description: Успешная обработка запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtResponse'
              examples:
                Ответ с токенами доступа и обновления:
                  description: Ответ с токенами доступа и обновления
                  value:
                    operationStatus: COMPLETED
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    accessToken: >-
                      eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJTdXBlck1hbjQiLCJleHAiOjE3NjAzMTA1NzcsInJvbGVzIjpbIlVTRVIiXX0.0jcf_wlE5r49nMnlvL5d_HFi0w2f-ssCsPlpcTWwUZTjbfCkzVi8KHmKPoemhoWijCSwnFqMu9RuPST2YC2NxA
                    refreshToken: >-
                      eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJTdXBlck1hbjQiLCJleHAiOjE3NjI5MDIyNzd9.wFx7TLzzErJ9unezD2Znu7JC0z-f_PQWuHmASsfgDGtCTcz98-TD_Ncvw5qeV_lwv3FrasiuBnByqFCArbBAuA
                    type: Bearer
                Ответ с токенами доступа:
                  description: Ответ с токенами доступа
                  value:
                    operationStatus: COMPLETED
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    accessToken: >-
                      eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJTdXBlck1hbjQiLCJleHAiOjE3NjAzMTA1NzcsInJvbGVzIjpbIlVTRVIiXX0.0jcf_wlE5r49nMnlvL5d_HFi0w2f-ssCsPlpcTWwUZTjbfCkzVi8KHmKPoemhoWijCSwnFqMu9RuPST2YC2NxA
                    type: Bearer
        '400':
          description: >-
            Запрос не соответствует заявленной схеме данных и не может быть
            обработан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке валидации:
                  description: Ответ с сообщением об ошибке валидации
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Нарушение формата данных
        '403':
          description: Обработка закончилась с ошибкой авторизации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке авторизации:
                  description: Ответ с сообщением об ошибке авторизации
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Неправильный пароль
        '404':
          description: Обработка закончилась с ошибкой поиска
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке поиска:
                  description: Ответ с сообщением об ошибке поиска
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Пользователь не найден
        '500':
          description: Обработка закончилась внутренней ошибкой сервиса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением о внутренней ошибкой сервиса:
                  description: Ответ с сообщением о внутренней ошибкой сервиса
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Внутренняя ошибка сервиса
      security:
        - JWT: []
  /api/v1/auth/login:
    post:
      tags:
        - Управление токенами доступа
      summary: Получение токена доступа и токена обновления
      operationId: login
      requestBody:
        description: >-
          Запрос на получение токена доступа и токена обновления по логину и
          паролю
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JwtRequest'
            examples:
              Запрос на получение токена доступа и токена обновления по логину и паролю:
                description: >-
                  Запрос на получение токена доступа и токена обновления по
                  логину и паролю
                value:
                  requestId: 123e4567-e89b-12d3-a456-426614174000
                  login: SuperMan
                  password: '**********'
        required: true
      responses:
        '200':
          description: Успешная обработка запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtResponse'
              examples:
                Ответ с токенами доступа и обновления:
                  description: Ответ с токенами доступа и обновления
                  value:
                    operationStatus: COMPLETED
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    accessToken: >-
                      eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJTdXBlck1hbjQiLCJleHAiOjE3NjAzMTA1NzcsInJvbGVzIjpbIlVTRVIiXX0.0jcf_wlE5r49nMnlvL5d_HFi0w2f-ssCsPlpcTWwUZTjbfCkzVi8KHmKPoemhoWijCSwnFqMu9RuPST2YC2NxA
                    refreshToken: >-
                      eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJTdXBlck1hbjQiLCJleHAiOjE3NjI5MDIyNzd9.wFx7TLzzErJ9unezD2Znu7JC0z-f_PQWuHmASsfgDGtCTcz98-TD_Ncvw5qeV_lwv3FrasiuBnByqFCArbBAuA
                    type: Bearer
                Ответ с токенами доступа:
                  description: Ответ с токенами доступа
                  value:
                    operationStatus: COMPLETED
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    accessToken: >-
                      eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJTdXBlck1hbjQiLCJleHAiOjE3NjAzMTA1NzcsInJvbGVzIjpbIlVTRVIiXX0.0jcf_wlE5r49nMnlvL5d_HFi0w2f-ssCsPlpcTWwUZTjbfCkzVi8KHmKPoemhoWijCSwnFqMu9RuPST2YC2NxA
                    type: Bearer
        '400':
          description: >-
            Запрос не соответствует заявленной схеме данных и не может быть
            обработан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке валидации:
                  description: Ответ с сообщением об ошибке валидации
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Нарушение формата данных
        '403':
          description: Обработка закончилась с ошибкой авторизации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке авторизации:
                  description: Ответ с сообщением об ошибке авторизации
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Неправильный пароль
        '404':
          description: Обработка закончилась с ошибкой поиска
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением об ошибке поиска:
                  description: Ответ с сообщением об ошибке поиска
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Пользователь не найден
        '500':
          description: Обработка закончилась внутренней ошибкой сервиса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                Ответ с сообщением о внутренней ошибкой сервиса:
                  description: Ответ с сообщением о внутренней ошибкой сервиса
                  value:
                    operationStatus: ERROR
                    responseId: 123e4567-e89b-12d3-a456-426614174000
                    errorMessage: Внутренняя ошибка сервиса
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        operationStatus:
          type: string
          enum:
            - COMPLETED
            - ERROR
        responseId:
          type: string
        errorMessage:
          type: string
...
    RefreshJwtRequest:
      type: object
      properties:
        requestId:
          type: string
          format: uuid
        refreshToken:
          type: string
          minLength: 1
      required:
        - refreshToken
        - requestId
    JwtResponse:
      type: object
      properties:
        operationStatus:
          type: string
          enum:
            - COMPLETED
            - ERROR
        responseId:
          type: string
        type:
          type: string
        accessToken:
          type: string
        refreshToken:
          type: string
    JwtRequest:
      type: object
      properties:
        requestId:
          type: string
          format: uuid
        login:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
      required:
        - login
        - password
        - requestId
  securitySchemes:
    JWT:
      type: http
      in: header
      scheme: bearer
      bearerFormat: 'Bearer '